Migration of all styles and completely new homepage
I just wanted to move the PostCSS in this ticket, but to move all CSS
more work on other tickets was needed. This ticket moves all CSS from
the old page and does some larger changes to the styles. Additionally I
introduced CSS Grid Layout as a base page grid, as well as many more
grids on component level. I designed and build a completely new homepage
and adjusted the styleguide.

* Migrate PostCSS styles to new website (closes #16)
* Use CSS Grid Layout Module Level 1 for page layout (closes #12)
* Adjust styleguide to mirror style of page (closes #14)
* Design completely new homepage (closes #6)
* Move Downloads to new project (closes #24)